fix(ui): Consolidate render/column logic
