+++ tavern/internal/www/src/components/tavern-base-ui/virtualized-table/index.tsx
@@ -1,11 +1,8 @@
 export { VirtualizedTable } from './VirtualizedTable';
-export { VirtualizedTableHeader } from './VirtualizedTableHeader';
-export { VirtualizedTableRow } from './VirtualizedTableRow';
 export { VirtualizedTableWrapper } from './VirtualizedTableWrapper';
 export type {
-    VirtualizedTableHeaderProps,
     VirtualizedTableProps,
     VirtualizedTableColumn,
-    VirtualizedTableRowProps,
     VirtualizedTableWrapperProps,
+    ExpandableConfig,
 } from './types';
